+++ src/main/webapp/WEB-INF/zdh/js/etl/etl_task_drools.js
@@ -105,9 +105,16 @@
 
           },
           'click #del': function (e, value, row, index) {
-              var ids = new Array();// 声明一个数组
-              ids.push(row.id)
-              deleteMs(ids)
+              layer.confirm('是否删除任务', {
+                  btn: ['确定','取消'] //按钮
+              }, function(index){
+                  var ids = new Array();// 声明一个数组
+                  ids.push(row.id)
+                  deleteMs(ids)
+              }, function(){
+
+              });
+
           }
       };
